+++ massage/src/components/Main/Main.js
@@ -1,0 +1,9 @@
+
+const Main = () => {
+
+   return (
+      <></>
+   )
+}
+
+export default Main
