+++ massage/src/components/Profile/Profile.js
@@ -1,0 +1,9 @@
+const Profile = () => {
+   return (
+      <div>
+         Profile
+      </div>
+   )
+}
+
+export default Profile
